// src/common/response/error-code.enum.ts
export enum ErrorCode {
  // Base
  INTERNAL_ERROR = 'INTERNAL_ERROR',

  // Auth
  EMAIL_TAKEN = 'EMAIL_TAKEN',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  REFRESH_TOKEN_REUSE = 'REFRESH_TOKEN_REUSE',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',

  // Catalog
  CATALOG_NOT_FOUND = 'CATALOG_NOT_FOUND',

  // Product
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  INSUFFICIENT_STOCK = 'INSUFFICIENT_STOCK',
}